add schema to pytest fixture
