+++ pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
 addopts = -xs --emoji --mypy-ini-file=tests/mypy/config.cfg
-testpaths=tests/
 DJANGO_SETTINGS_MODULE=tests.django.django_settings
+testpaths=tests/ strawberry/flask/tests/
