+++ tests/conftest.py
@@ -1,2 +1,3 @@
 def pytest_emoji_xfailed(config):
     return "🤷‍♂️ ", "XFAIL 🤷‍♂️ "
+
